requirements.txt
apispec==6.6.1
bcrypt==4.0.1
blinker==1.8.2
click==8.1.8
colorama==0.4.6
coverage==7.6.1
Deprecated==1.3.1
exceptiongroup==1.3.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==5.0.0
rich==13.9.4
SQLAlchemy==2.0.45
tomli==2.4.0
typing_extensions==4.13.2
webargs==8.6.0
Werkzeug==3.0.6
wrapt==2.0.1
zipp==3.20.2
